[CBR-227] Remove legacy wallet layer
This was never used, and will never be used. That is unfortunate, but at this
point won't change, and having the legacy layer makes it more awkward to modify
the layer since we have to make modifications to a layer that is never actually
used.
