+++ wallet-new/test/unit/Test/Spec/Fixture.hs
@@ -27,7 +27,7 @@ import qualified Cardano.Wallet.Kernel.Keystore as Keystore
 import           Cardano.Wallet.Kernel.NodeStateAdaptor (nodeStateUnavailable)
 import           Cardano.Wallet.WalletLayer (ActiveWalletLayer,
                      PassiveWalletLayer)
-import qualified Cardano.Wallet.WalletLayer as WalletLayer
+import qualified Cardano.Wallet.WalletLayer.Kernel as WalletLayer.Kernel
 
 -- | Do not pollute the test runner output with logs.
 devNull :: Severity -> Text -> IO ()
@@ -42,7 +42,7 @@ withLayer :: MonadIO m
           -> PropertyM IO a
 withLayer cc = do
     liftIO $ Keystore.bracketTestKeystore $ \keystore -> do
-        WalletLayer.bracketKernelPassiveWallet devNull keystore nodeStateUnavailable $ \layer wallet -> do
+        WalletLayer.Kernel.bracketPassiveWallet devNull keystore nodeStateUnavailable $ \layer wallet -> do
             cc layer wallet
 
 type GenPassiveWalletFixture x = PropertyM IO (PassiveWallet -> IO x)
@@ -55,7 +55,7 @@ withPassiveWalletFixture :: MonadIO m
 withPassiveWalletFixture prepareFixtures cc = do
     generateFixtures <- prepareFixtures
     liftIO $ Keystore.bracketTestKeystore $ \keystore -> do
-        WalletLayer.bracketKernelPassiveWallet devNull keystore nodeStateUnavailable $ \layer wallet -> do
+        WalletLayer.Kernel.bracketPassiveWallet devNull keystore nodeStateUnavailable $ \layer wallet -> do
             fixtures <- generateFixtures wallet
             cc keystore layer wallet fixtures
 
@@ -66,9 +66,9 @@ withActiveWalletFixture :: MonadIO m
 withActiveWalletFixture prepareFixtures cc = do
     generateFixtures <- prepareFixtures
     liftIO $ Keystore.bracketTestKeystore $ \keystore -> do
-        WalletLayer.bracketKernelPassiveWallet devNull keystore nodeStateUnavailable $ \passiveLayer passiveWallet -> do
+        WalletLayer.Kernel.bracketPassiveWallet devNull keystore nodeStateUnavailable $ \passiveLayer passiveWallet -> do
             withDefConfiguration $ \pm -> do
-                WalletLayer.bracketKernelActiveWallet pm passiveLayer passiveWallet diffusion $ \activeLayer activeWallet -> do
+                WalletLayer.Kernel.bracketActiveWallet pm passiveLayer passiveWallet diffusion $ \activeLayer activeWallet -> do
                     fixtures <- generateFixtures keystore activeWallet
                     cc keystore activeLayer activeWallet fixtures
     where
